refactor(src): rename interface folder
