+++ src/auth/guard/authorization/roles.guard.ts
@@ -3,8 +3,8 @@ import { Reflector } from "@nestjs/core";
 import { ServiceException } from "../../../_common/filter/exception/service/serviceException";
 import { ROLES_KEY } from "../../../user/decorator/role";
 import { UserRole } from "../../../user/entity/user.entity";
-import { AUTH_GUARD_PAYLOAD } from "../type/requestPayload";
-import { AuthGuardRequest } from "../type/AuthRequest";
+import { AUTH_GUARD_PAYLOAD } from "../types/requestPayload";
+import { AuthGuardRequest } from "../types/AuthRequest";
 
 @Injectable()
 export class RolesGuard implements CanActivate {
