+++ src/painting/types/shortPainting.ts
