+++ src/painting/types/weeklyArtWorkSet.ts
