+++ src/quiz/types/relatedPaintingIds.ts
