+++ src/quiz/types/relatedPaintings.ts
